split AWS functionalities into a subfolder and to be used as an extension
