+++ tests/aws/test_aws_safe.py
@@ -2,7 +2,8 @@ import os
 
 import pytest
 
-from sentinelhub import AwsConstants, AwsProductRequest, AwsTileRequest, DataCollection, read_data, write_data
+from sentinelhub import DataCollection, read_data, write_data
+from sentinelhub.aws import AwsConstants, AwsProductRequest, AwsTileRequest
 
 
 @pytest.fixture(name="safe_folder", scope="session")
